WIP: Add parsing options
It must be possible to choose how the parsing is made (eg: based on RFC 4180 or on an extension that supports any unicode character).
This commit makes use of the new ParsingOptions type, which currently only has the RFC 4180 option. Note that the parser hasn't been tested extensively enough to say that it closely matches the specifications of RFC 4180.